+++ CHANGELOG.md
@@ -10,10 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New logger for CI, like Travis and AppVeyor, with colored output.
+- ounit-lwt to build test with OUnit and Lwt. It also allows
+  to use the runner "processes" to run test in parallel. (Closes: OF#1765)
 
 ### Changed
 - assert_command only displays the difference with the initial environment.
   This avoids to have hundreeds of lines of useless environment variables.
+- Upgrade minimal OCaml version to 4.02, since dune requires at least this
+  version.
 
 ### Fixed
 - Run garbage collection in between tests to prevent unexpected bugs in GC (e.g